chore: move to lexical folder
