+++ src/components/Shared/Lexical/Plugins/AtMentionsPlugin.tsx
@@ -17,7 +17,7 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 import * as ReactDOM from 'react-dom';
 import { AVATAR } from 'src/constants';
 
-import { $createMentionNode } from './MentionsNode';
+import { $createMentionNode } from '../Nodes/MentionsNode';
 
 const PUNCTUATION = '\\.,\\+\\*\\?\\$\\@\\|#{}\\(\\)\\^\\-\\[\\]\\\\/!%\'"~=<>_:;';
 const NAME = '\\b[A-Z][^\\s' + PUNCTUATION + ']';
